# docs/openapi/tiktok-open-api.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TikTok Open API",
    "version": "v1",
    "description": "Implemented external API for TikTok authorization, upload, publish, and query. All returned resource IDs are opaque and every protected resource is isolated by the authenticated clientId. Administrative management APIs are not part of this contract."
  },
  "servers": [{"url": "https://tkassetplant.fnn.net.cn/admin-api/tk/open/v1/tiktok", "description": "Production"}],
  "tags": [{"name": "Authorization"}, {"name": "Media"}, {"name": "Publish"}],
  "security": [{"ClientId": [], "Timestamp": [], "Nonce": [], "Signature": []}],
  "paths": {
    "/auth/sessions": {
      "post": {"tags": ["Authorization"], "summary": "Create an authorization session", "parameters": [{"$ref": "#/components/parameters/RequestId"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuthSessionCreate"}}}}, "responses": {"200": {"description": "Session created", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuthSessionEnvelope"}}}}, "400": {"$ref": "#/components/responses/Error400"}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "413": {"$ref": "#/components/responses/Error413"}, "429": {"$ref": "#/components/responses/Error429"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}}
    },
    "/auth/sessions/{authSessionId}": {
      "get": {"tags": ["Authorization"], "summary": "Get authorization session status", "parameters": [{"$ref": "#/components/parameters/AuthSessionId"}, {"$ref": "#/components/parameters/RequestId"}], "responses": {"200": {"description": "Session status", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AuthStatusEnvelope"}}}}, "400": {"$ref": "#/components/responses/Error400"}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "429": {"$ref": "#/components/responses/Error429"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}}
    },
    "/connections": {
      "get": {"tags": ["Authorization"], "summary": "List account connections", "parameters": [{"name": "externalAccountId", "in": "query", "schema": {"type": "string", "maxLength": 128}}, {"name": "status", "in": "query", "schema": {"type": "string"}}, {"$ref": "#/components/parameters/RequestId"}], "responses": {"200": {"description": "Connections", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConnectionsEnvelope"}}}}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "429": {"$ref": "#/components/responses/Error429"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}}
    },
    "/connections/{connectionId}/disconnect": {
      "post": {"tags": ["Authorization"], "summary": "Disconnect an account", "parameters": [{"$ref": "#/components/parameters/ConnectionId"}, {"$ref": "#/components/parameters/RequestId"}], "responses": {"200": {"description": "Disconnected", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BooleanEnvelope"}}}}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "429": {"$ref": "#/components/responses/Error429"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}}
    },
    "/auth/callback": {
      "get": {"tags": ["Authorization"], "summary": "TikTok OAuth callback", "description": "TikTok redirect target. This endpoint skips request HMAC and validates OAuth state. It is not an external caller operation.", "security": [], "parameters": [{"name": "code", "in": "query", "schema": {"type": "string"}}, {"name": "state", "in": "query", "schema": {"type": "string"}}, {"name": "error", "in": "query", "schema": {"type": "string"}}, {"name": "error_description", "in": "query", "schema": {"type": "string"}}], "responses": {"200": {"description": "HTML authorization outcome", "content": {"text/html": {"schema": {"type": "string"}}}}}}
    },
    "/media/uploads": {
      "post": {"tags": ["Media"], "summary": "Create an upload session", "parameters": [{"$ref": "#/components/parameters/RequestId"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UploadCreate"}}}}, "responses": {"200": {"description": "Upload session", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UploadSessionEnvelope"}}}}, "400": {"$ref": "#/components/responses/Error400"}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "413": {"$ref": "#/components/responses/Error413"}, "429": {"$ref": "#/components/responses/Error429"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}}
    },
    "/media/uploads/{uploadId}": {
      "get": {"tags": ["Media"], "summary": "Get upload status", "parameters": [{"$ref": "#/components/parameters/UploadId"}, {"$ref": "#/components/parameters/RequestId"}], "responses": {"200": {"description": "Upload status", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UploadStatusEnvelope"}}}}, "400": {"$ref": "#/components/responses/Error400"}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "429": {"$ref": "#/components/responses/Error429"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}},
      "delete": {"tags": ["Media"], "summary": "Cancel an upload session", "parameters": [{"$ref": "#/components/parameters/UploadId"}, {"$ref": "#/components/parameters/RequestId"}], "responses": {"200": {"description": "Cancelled", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BooleanEnvelope"}}}}, "400": {"$ref": "#/components/responses/Error400"}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "429": {"$ref": "#/components/responses/Error429"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}}
    },
    "/media/uploads/{uploadId}/chunks/{chunkIndex}": {
      "put": {"tags": ["Media"], "summary": "Upload a LOCAL-mode binary chunk", "description": "Available only when create-upload returned uploadMode LOCAL. chunkIndex is zero-based. Use the returned chunkSize; only the final chunk may be smaller.", "parameters": [{"$ref": "#/components/parameters/UploadId"}, {"name": "chunkIndex", "in": "path", "required": true, "schema": {"type": "integer", "minimum": 0}}, {"$ref": "#/components/parameters/RequestId"}], "requestBody": {"required": true, "content": {"application/octet-stream": {"schema": {"type": "string", "format": "binary"}}}}, "responses": {"200": {"description": "Chunk stored", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BooleanEnvelope"}}}}, "400": {"$ref": "#/components/responses/Error400"}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "413": {"$ref": "#/components/responses/Error413"}, "429": {"$ref": "#/components/responses/Error429"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}}
    },
    "/media/uploads/{uploadId}/complete": {
      "post": {"tags": ["Media"], "summary": "Complete an upload", "description": "For OSS mode, the service HEAD-checks the object length and SHA-256 metadata before returning READY media.", "parameters": [{"$ref": "#/components/parameters/UploadId"}, {"$ref": "#/components/parameters/RequestId"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UploadComplete"}}}}, "responses": {"200": {"description": "Ready media", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MediaEnvelope"}}}}, "400": {"$ref": "#/components/responses/Error400"}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "413": {"$ref": "#/components/responses/Error413"}, "429": {"$ref": "#/components/responses/Error429"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}}
    },
    "/publish/tasks": {
      "post": {"tags": ["Publish"], "summary": "Create an asynchronous publish task", "parameters": [{"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 128}}, {"$ref": "#/components/parameters/RequestId"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PublishCreate"}}}}, "responses": {"200": {"description": "Task created or idempotent result", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TaskEnvelope"}}}}, "400": {"$ref": "#/components/responses/Error400"}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "409": {"$ref": "#/components/responses/Error409"}, "413": {"$ref": "#/components/responses/Error413"}, "429": {"$ref": "#/components/responses/Error429"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}}
    },
    "/publish/tasks/{taskId}": {
      "get": {"tags": ["Publish"], "summary": "Get publish task", "parameters": [{"$ref": "#/components/parameters/TaskId"}, {"$ref": "#/components/parameters/RequestId"}], "responses": {"200": {"description": "Task", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TaskEnvelope"}}}}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "429": {"$ref": "#/components/responses/Error429"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}}
    },
    "/publish/tasks/{taskId}/details": {
      "get": {"tags": ["Publish"], "summary": "List publish task details", "parameters": [{"$ref": "#/components/parameters/TaskId"}, {"$ref": "#/components/parameters/RequestId"}], "responses": {"200": {"description": "Per-connection details", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DetailsEnvelope"}}}}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "429": {"$ref": "#/components/responses/Error429"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}}
    },
    "/publish/details/{detailId}/retry": {
      "post": {"tags": ["Publish"], "summary": "Retry a failed publish detail", "parameters": [{"$ref": "#/components/parameters/DetailId"}, {"$ref": "#/components/parameters/RequestId"}], "responses": {"200": {"description": "Retry queued", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/BooleanEnvelope"}}}}, "400": {"$ref": "#/components/responses/Error400"}, "401": {"$ref": "#/components/responses/Error401"}, "403": {"$ref": "#/components/responses/Error403"}, "404": {"$ref": "#/components/responses/Error404"}, "429": {"$ref": "#/components/responses/Error429"}, "500": {"$ref": "#/components/responses/Error500"}, "503": {"$ref": "#/components/responses/Error503"}}}
    }
  },
  "components": {
    "securitySchemes": {
      "ClientId": {"type": "apiKey", "in": "header", "name": "X-TK-Client-Id", "description": "Issued client ID."},
      "Timestamp": {"type": "apiKey", "in": "header", "name": "X-TK-Timestamp", "description": "Unix epoch seconds, within 300 seconds."},
      "Nonce": {"type": "apiKey", "in": "header", "name": "X-TK-Nonce", "description": "Unique anti-replay nonce."},
      "Signature": {"type": "apiKey", "in": "header", "name": "X-TK-Signature", "description": "Base64 HMAC-SHA256 of METHOD, request target, timestamp, nonce, and SHA-256 body hash."}
    },
    "parameters": {
      "RequestId": {"name": "X-TK-Request-Id", "in": "header", "required": false, "schema": {"type": "string", "pattern": "^[A-Za-z0-9._:-]{1,128}$"}, "description": "Caller trace ID. The server generates one when omitted or invalid."},
      "AuthSessionId": {"name": "authSessionId", "in": "path", "required": true, "schema": {"type": "string"}},
      "ConnectionId": {"name": "connectionId", "in": "path", "required": true, "schema": {"type": "string"}},
      "UploadId": {"name": "uploadId", "in": "path", "required": true, "schema": {"type": "string"}},
      "TaskId": {"name": "taskId", "in": "path", "required": true, "schema": {"type": "string"}},
      "DetailId": {"name": "detailId", "in": "path", "required": true, "schema": {"type": "string"}}
    },
    "responses": {
      "Error400": {"description": "Parameter or resource-state error. Codes include OPEN_API_PARAMETER_INVALID, MEDIA_FILE_INVALID, MEDIA_NOT_READY, MEDIA_UPLOAD_EXPIRED, MEDIA_UPLOAD_STATUS_INVALID (chunk, complete, and cancel require UPLOADING), UPLOAD_MODE_INVALID, IDEMPOTENCY_KEY_REQUIRED, IDEMPOTENCY_KEY_INVALID, CONNECTION_NOT_AUTHORIZED, and PUBLISH_RETRY_STATUS_INVALID.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error401": {"description": "Authentication error. Codes include OPEN_API_AUTH_HEADER_MISSING, OPEN_API_CLIENT_INVALID, OPEN_API_SIGNATURE_INVALID, OPEN_API_TIMESTAMP_EXPIRED, and OPEN_API_NONCE_REPLAYED.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error403": {"description": "Client policy error: OPEN_API_IP_NOT_ALLOWED or OPEN_API_PERMISSION_DENIED.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error404": {"description": "Opaque resource not found. Codes include AUTH_SESSION_NOT_FOUND, CONNECTION_NOT_FOUND, MEDIA_NOT_FOUND, PUBLISH_TASK_NOT_FOUND, and PUBLISH_DETAIL_NOT_FOUND.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error409": {"description": "Idempotency conflict: IDEMPOTENCY_KEY_CONFLICT.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error413": {"description": "Request body or media too large. Codes include OPEN_API_BODY_TOO_LARGE and MEDIA_FILE_TOO_LARGE.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error429": {"description": "Rate/quota error: OPEN_API_RATE_LIMITED or OPEN_API_QUOTA_EXCEEDED.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error500": {"description": "Internal error: OPEN_API_INTERNAL_ERROR.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}},
      "Error503": {"description": "Temporary dependency/configuration error. Codes include TIKTOK_CONFIG_REQUIRED, AUTHORIZATION_FAILED, OSS_CONFIG_REQUIRED, MEDIA_UPLOAD_FAILED, IDEMPOTENCY_RESULT_UNAVAILABLE, OPEN_API_GUARD_UNAVAILABLE, and OPEN_API_SECRET_UNAVAILABLE.", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}
    },
    "schemas": {
      "Envelope": {"type": "object", "required": ["code", "msg", "data", "requestId"], "properties": {"code": {"oneOf": [{"type": "integer", "enum": [0]}, {"type": "string"}]}, "msg": {"type": "string"}, "data": {"nullable": true}, "requestId": {"type": "string"}}},
      "ErrorEnvelope": {"allOf": [{"$ref": "#/components/schemas/Envelope"}], "example": {"code": "MEDIA_NOT_READY", "msg": "media is not ready", "data": null, "requestId": "req_example"}},
      "AuthSessionCreate": {"type": "object", "required": ["externalAccountId", "authMode"], "properties": {"externalAccountId": {"type": "string", "maxLength": 128}, "authMode": {"type": "string", "enum": ["REDIRECT", "QR_CODE"]}, "clientState": {"type": "string", "maxLength": 512}}},
      "AuthSession": {"type": "object", "properties": {"authSessionId": {"type": "string"}, "externalAccountId": {"type": "string"}, "clientState": {"type": "string", "nullable": true}, "authMode": {"type": "string", "enum": ["REDIRECT", "QR_CODE"]}, "authorizeUrl": {"type": "string", "format": "uri", "nullable": true}, "qrcodeUrl": {"type": "string", "format": "uri", "nullable": true}, "status": {"type": "string", "enum": ["WAITING", "SUCCESS", "FAILED", "EXPIRED"]}, "expireTime": {"type": "string", "format": "date-time"}}},
      "AuthStatus": {"type": "object", "properties": {"authSessionId": {"type": "string"}, "externalAccountId": {"type": "string"}, "clientState": {"type": "string", "nullable": true}, "connectionId": {"type": "string", "nullable": true}, "accountName": {"type": "string", "nullable": true}, "status": {"type": "string", "enum": ["WAITING", "SUCCESS", "FAILED", "EXPIRED"]}, "failReason": {"type": "string", "nullable": true}, "expireTime": {"type": "string", "format": "date-time"}}},
      "Connection": {"type": "object", "properties": {"connectionId": {"type": "string"}, "externalAccountId": {"type": "string"}, "accountName": {"type": "string", "nullable": true}, "username": {"type": "string", "nullable": true}, "avatarUrl": {"type": "string", "format": "uri", "nullable": true}, "authStatus": {"type": "string"}, "tokenStatus": {"type": "string"}, "lastAuthTime": {"type": "string", "format": "date-time", "nullable": true}}},
      "UploadCreate": {"type": "object", "required": ["fileName", "fileSize"], "properties": {"fileName": {"type": "string", "pattern": "^.+\\.([mM][pP]4|[mM][oO][vV]|[wW][eE][bB][mM])$"}, "fileSize": {"type": "integer", "format": "int64", "minimum": 1, "maximum": 1000000000}, "contentType": {"type": "string"}, "sha256": {"type": "string", "description": "Optional full-file SHA-256 value. When supplied, complete must use the same value and storage metadata/file content is validated."}}},
      "OssFields": {"type": "object", "properties": {"policy": {"type": "string"}, "ossAccessKeyId": {"type": "string"}, "signature": {"type": "string"}, "xOssMetaSha256": {"type": "string", "nullable": true}}},
      "UploadSession": {"type": "object", "properties": {"uploadId": {"type": "string"}, "uploadMode": {"type": "string", "enum": ["OSS", "LOCAL"]}, "chunkSize": {"type": "integer", "nullable": true}, "totalChunks": {"type": "integer", "nullable": true}, "uploadUrl": {"type": "string", "format": "uri", "nullable": true}, "objectKey": {"type": "string", "nullable": true}, "fields": {"nullable": true, "allOf": [{"$ref": "#/components/schemas/OssFields"}]}, "expireTime": {"type": "string", "format": "date-time"}}},
      "UploadStatus": {"type": "object", "properties": {"uploadId": {"type": "string"}, "mediaId": {"type": "string", "nullable": true}, "uploadMode": {"type": "string", "enum": ["OSS", "LOCAL"]}, "fileSize": {"type": "integer", "format": "int64"}, "uploadedSize": {"type": "integer", "format": "int64"}, "status": {"type": "string", "enum": ["UPLOADING", "READY", "CANCELLED"]}}},
      "UploadComplete": {"type": "object", "required": ["fileSize"], "properties": {"fileSize": {"type": "integer", "format": "int64", "minimum": 1}, "sha256": {"type": "string"}, "coverTimestampMs": {"type": "integer", "format": "int64"}}},
      "Media": {"type": "object", "properties": {"mediaId": {"type": "string"}, "uploadId": {"type": "string"}, "fileName": {"type": "string"}, "fileSize": {"type": "integer", "format": "int64"}, "contentType": {"type": "string"}, "status": {"type": "string", "enum": ["READY"]}}},
      "PublishCreate": {"type": "object", "required": ["connectionIds", "mediaId", "postMode", "privacyLevel"], "properties": {"connectionIds": {"type": "array", "minItems": 1, "maxItems": 20, "items": {"type": "string"}}, "mediaId": {"type": "string"}, "title": {"type": "string", "maxLength": 512}, "caption": {"type": "string", "maxLength": 2200}, "postMode": {"type": "string", "enum": ["DIRECT_POST", "UPLOAD_TO_INBOX"]}, "privacyLevel": {"type": "string"}, "allowComment": {"type": "boolean", "default": true}, "allowDuet": {"type": "boolean", "default": false}, "allowStitch": {"type": "boolean", "default": false}, "commercialContent": {"type": "boolean", "default": false}, "brandContent": {"type": "boolean", "default": false}, "aigcContent": {"type": "boolean", "default": true}, "externalRequestId": {"type": "string", "maxLength": 128}}},
      "Task": {"type": "object", "properties": {"taskId": {"type": "string"}, "mediaId": {"type": "string"}, "externalRequestId": {"type": "string", "nullable": true}, "status": {"type": "string", "enum": ["PENDING", "PROCESSING", "SUCCESS", "FAILED", "PARTIAL_SUCCESS"]}, "accountCount": {"type": "integer"}, "successCount": {"type": "integer"}, "failedCount": {"type": "integer"}, "pendingCount": {"type": "integer"}, "failReason": {"type": "string", "nullable": true}, "createTime": {"type": "string", "format": "date-time"}, "updateTime": {"type": "string", "format": "date-time"}}},
      "PublishDetail": {"type": "object", "properties": {"detailId": {"type": "string"}, "taskId": {"type": "string"}, "connectionId": {"type": "string"}, "accountName": {"type": "string", "nullable": true}, "status": {"type": "string"}, "tiktokStatus": {"type": "string"}, "publishId": {"type": "string", "nullable": true}, "publishUrl": {"type": "string", "format": "uri", "nullable": true}, "failReason": {"type": "string", "nullable": true}, "retryCount": {"type": "integer"}, "updateTime": {"type": "string", "format": "date-time"}}},
      "BooleanEnvelope": {"allOf": [{"$ref": "#/components/schemas/Envelope"}], "example": {"code": 0, "msg": "OK", "data": true, "requestId": "req_example"}},
      "AuthSessionEnvelope": {"allOf": [{"$ref": "#/components/schemas/Envelope"}], "properties": {"data": {"$ref": "#/components/schemas/AuthSession"}}},
      "AuthStatusEnvelope": {"allOf": [{"$ref": "#/components/schemas/Envelope"}], "properties": {"data": {"$ref": "#/components/schemas/AuthStatus"}}},
      "ConnectionsEnvelope": {"allOf": [{"$ref": "#/components/schemas/Envelope"}], "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/Connection"}}}},
      "UploadSessionEnvelope": {"allOf": [{"$ref": "#/components/schemas/Envelope"}], "properties": {"data": {"$ref": "#/components/schemas/UploadSession"}}},
      "UploadStatusEnvelope": {"allOf": [{"$ref": "#/components/schemas/Envelope"}], "properties": {"data": {"$ref": "#/components/schemas/UploadStatus"}}},
      "MediaEnvelope": {"allOf": [{"$ref": "#/components/schemas/Envelope"}], "properties": {"data": {"$ref": "#/components/schemas/Media"}}},
      "TaskEnvelope": {"allOf": [{"$ref": "#/components/schemas/Envelope"}], "properties": {"data": {"$ref": "#/components/schemas/Task"}}},
      "DetailsEnvelope": {"allOf": [{"$ref": "#/components/schemas/Envelope"}], "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/PublishDetail"}}}}
    }
  }
}